feat: Organize mobile navigation
